Various changes for now, use test set as validation. Support for multiple models.
Validation edit distance will now correspond to test edit distance. This is just for testing and will be changed later. Because much of the code is likely to overlap between new models and the basic model, have CTCModel inherit from Model. Add run.py for running the model (move this functionality out of basic_model.py).